added tests to improve code covarege
