+++ src/test/java/com/microsoft/sqlserver/jdbc/TestUtils.java
@@ -21,6 +21,8 @@ import java.sql.Timestamp;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Calendar;
+import java.util.Locale;
+import java.util.ResourceBundle;
 
 import com.microsoft.sqlserver.testframework.AbstractSQLGenerator;
 import com.microsoft.sqlserver.testframework.Constants;
@@ -71,52 +73,30 @@ public class TestUtils {
     private static Boolean isAzureDW = null;
     private static Boolean isAzureMI = null;
 
-    /*
-     * SERVERPROPERTY('EngineEdition') can be used to determine whether the db server is SQL Azure.
-     * It should return 6 for SQL Azure DW. This is more reliable than @@version or
-     * serverproperty('edition').
-     * Reference: http://msdn.microsoft.com/en-us/library/ee336261.aspx
+    /**
+     * Check if connected to Azure server.
      * 
-     * SERVERPROPERTY('EngineEdition') means
-     * Database Engine edition of the instance of SQL Server installed on the server.
-     * 1 = Personal or Desktop Engine (Not available for SQL Server.)
-     * 2 = Standard (This is returned for Standard and Workgroup.)
-     * 3 = Enterprise (This is returned for Enterprise, Enterprise Evaluation, and Developer.)
-     * 4 = Express (This is returned for Express, Express with Advanced Services, and Windows Embedded SQL.)
-     * 5 = SQL Azure
-     * 6 = SQL Azure DW
-     * 8 = Managed Instance
-     * Base data type: int
+     * @see com.microsoft.sqlserver.jdbc.SQLServerConnection#isAzure()
      */
     public static boolean isAzure(Connection con) {
-        if (null == isAzure) {
-            try (Statement stmt = con.createStatement();
-                    ResultSet rs = stmt.executeQuery("SELECT CAST(SERVERPROPERTY('EngineEdition') as INT)")) {
-                rs.next();
-
-                int engineEdition = rs.getInt(1);
-                isAzure = (engineEdition == ENGINE_EDITION_FOR_SQL_AZURE
-                        || engineEdition == ENGINE_EDITION_FOR_SQL_AZURE_DW
-                        || engineEdition == ENGINE_EDITION_FOR_SQL_AZURE_MI);
-                isAzureDW = (engineEdition == ENGINE_EDITION_FOR_SQL_AZURE_DW);
-                isAzureMI = (engineEdition == ENGINE_EDITION_FOR_SQL_AZURE_MI);
-
-            } catch (SQLException e) {
-                isAzure = false;
-                isAzureDW = false;
-                isAzureMI = false;
-            }
-            return isAzure;
-        } else {
-            return isAzure;
-        }
+        return ((SQLServerConnection) con).isAzure();
     }
 
+    /**
+     * Check if connection is to Azure DW server.
+     * 
+     * @see com.microsoft.sqlserver.jdbc.SQLServerConnection#isAzureDW()
+     */
     public static boolean isAzureDW(Connection con) {
         isAzure(con);
         return isAzureDW;
     }
 
+    /**
+     * Check if connection is to Azure MI server.
+     * 
+     * @see com.microsoft.sqlserver.jdbc.SQLServerConnection#isAzureMI()
+     */
     public static boolean isAzureMI(Connection con) {
         isAzure(con);
         return isAzureMI;
@@ -701,4 +681,26 @@ public class TestUtils {
     public static String escapeSingleQuotes(String name) {
         return name.replace("'", "''");
     }
+
+    public static final ResourceBundle rBundle = getDefaultLocaleBundle();
+
+    /**
+     * Returns the root bundle. This is the bundle from SQLServerResource.java - the English version that gets updated
+     * in development process.
+     *
+     * @return root bundle.
+     */
+    private static ResourceBundle getDefaultLocaleBundle() {
+        return ResourceBundle.getBundle("com.microsoft.sqlserver.jdbc.SQLServerResource", Locale.getDefault());
+    }
+
+    /**
+     * Creates a regex where all '{#}' fields will return true for any value when calling match
+     *
+     * @return regex expression.
+     */
+    public static String formatErrorMsg(String s) {
+        return (".*\\Q" + TestUtils.rBundle.getString(s) + "\\E").replaceAll("\\{+[0-9]+\\}", "\\\\E.*\\\\Q");
+    }
+
 }
